requirements.txt
SQLAlchemy~=2.0.28
fastapi~=0.110.0
pydantic~=2.6.3
python-dotenv==1.0.1
uvicorn
psycopg2
